\name{cpt-package}
\alias{cpt-package}
\docType{package}

\title{
Classification Permutation Test
}

\description{
Description: Non-parametric test for equality of multivariate distributions.  Trains a classifier to classify (multivariate) observations as coming from one of several distributions.  If the classifier is able to classify the observations better than would be expected by chance (using permutation inference), then the null hypothesis that the distributions are equal is rejected.  
}

\details{
\tabular{ll}{
Package: \tab cpt\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-2-26\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}

\keyword{multivariate} 

\seealso{ 
\code{\link{cpt}}
} 

\author{ 
  Johann Gagnon-Bartsch <johanngb@umich.edu> 
} 

\references{
The Classification Permutation Test:  A Nonparametric Test for Equality of Multivariate Distributions.  Gagnon-Bartsch and Shem-Tov, 2016.  Available at: \href{https://arxiv.org/abs/1611.06408}{https://arxiv.org/abs/1611.06408}.
}

